module2-task1 + refactoring
