+++ src/presenter.js
@@ -1,24 +1,22 @@
-import { render } from './render';
-import CreationFormView from './view/creationForm-view';
-import EditingFormView from './view/editingForm-view';
-import ListView from './view/list-view';
-import PointView from './view/point-view';
-import SortView from './view/sort-view';
+import TripListView from '../view/trip-list-view';
+import TripPointView from '../view/trip-point-view';
+import TripPointCreatorView from '../view/trip-point-creator-view';
+import TripPointEditorView from '../view/trip-point-editor-view';
+import { render } from '../render';
 
 export default class PointListPresenter {
-  pointList = new ListView();
+  pointList = new TripListView();
 
-  constructor({pointListContainer}) {
+  init(pointListContainer, pointsModel) {
     this.pointListContainer = pointListContainer;
-  }
+    this.pointsModel = pointsModel;
+    this.pointListPoints = [...this.pointsModel.getTasks()];
 
-  init() {
-    render(new SortView, this.pointListContainer);
     render(this.pointList, this.pointListContainer);
-    render(new EditingFormView, this.pointList.getElement());
-    render(new CreationFormView, this.pointList.getElement());
-    for (let i = 0; i < 3; i++) {
-      render(new PointView, this.pointList.getElement());
+    render(new TripPointEditorView(this.pointListPoints[0]), this.pointList.getElement());
+    render(new TripPointCreatorView, this.pointList.getElement());
+    for (let i = 0; i < this.pointListPoints.length; i++) {
+      render(new TripPointView(this.pointListPoints[i]), this.pointList.getElement());
     }
   }
 }
